# Stress Testing API/Statistics with K6 Performance Benchmarking