--- a/pantau_tular/settings.py
+++ b/pantau_tular/settings.py
@@ -44,12 +44,16 @@ sentry_sdk.init(
     # Set traces_sample_rate to 1.0 to capture 100%
     # of transactions for tracing.
     traces_sample_rate=1.0,
+    # Integrasi Django
+    integrations=[DjangoIntegration()],
     # Set profile_session_sample_rate to 1.0 to profile 100%
     # of profile sessions.
     profile_session_sample_rate=1.0,
     # Set profile_lifecycle to "trace" to automatically
     # run the profiler on when there is an active transaction
-    profile_lifecycle="trace"
+    profile_lifecycle="trace",
+    # Mengirim informasi user agar mudah melacak siapa yang mengalami error
+    send_default_pii=True
 )
 
 
@@ -82,6 +86,7 @@ MIDDLEWARE = [
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
     'django_prometheus.middleware.PrometheusAfterMiddleware',
+    'pt_backend.middleware.SentryContextMiddleware',
 ]
 
 REST_FRAMEWORK = {