@KenBalya 2025-05-19
💬 Comments & Recommendations
Strengths
- Comprehensive Load Simulation: The test spans a realistic range of concurrent users (10–150), capturing both normal and peak usage scenarios effectively.

- Filter Diversity: Including both simple and complex filters reflects actual usage patterns and helps pinpoint performance weaknesses related to query complexity.

- Error Handling & Metrics: Excellent inclusion of success/error rates, latency percentiles, and dynamic adaptation to load (smart sleep times, filter toggling).

- Detailed Bottleneck Insights: The clear identification of server processing time as the main bottleneck gives actionable insight for backend optimization.

⚠️ Improvement Suggestions
1. High Latency Beyond 100 Users
Issue: At >100 concurrent users, average response time nears 6s and 95th percentile exceeds 8s.

Action:

Consider backend optimization (e.g., database query indexing, query caching, async processing for filters).

Investigate introducing load shedding or queuing mechanisms for non-critical requests under extreme load.

2. Complex Filter Processing Time
Issue: Filters combining date ranges, disease, and portal are disproportionately slow.

Action:

- Profile the query plans generated by these combinations.

- Apply materialized views or pre-aggregated statistics for frequent combinations to offload compute cost.

3. Timeout Errors Over 100 Users
Issue: Timeout errors appear under high concurrency.

Action:

Validate server timeout thresholds vs client wait time.

Consider introducing rate-limiting feedback to inform clients of load status.

Tune server thread pool / worker limits if applicable.

4. Response Time Target
Observation: 95% response threshold is currently 5s; however, 6,845ms at P95 shows it isn’t met under full load.

Suggestion:

Adjust SLA or invest in endpoint partitioning (e.g., async endpoints or paginated/stateless APIs) for better UX at scale.

5. Test Repeatability & Automation
Consider integrating the k6 tests into CI/CD pipeline for regression benchmarking.

Add multiple test iterations with warmup/cooldown phases to smooth out variability and identify persistent patterns.

📌 Final Thoughts
While the system performs well under moderate load, stress beyond 100 users exposes latency and occasional stability issues. Prioritizing backend optimization and investing in scalable architectural patterns will be critical for supporting growth. This is an excellent foundation for iterative tuning and production-hardening.